+++ Runtime/Scripts/UninterruptedDeferAgent.cs
@@ -2,8 +2,6 @@
 
     public class UninterruptedDeferAgent : IDeferAgent
     {
-        public void Reset() {}
-
         public bool ShouldDefer() {
             return false;
         }